refactor(presentation): VoiceNoteViewModel 프리뷰 확장을 별도 파일로 분리
